# swagger.yaml
openapi: 3.0.0
info:
  title: School Medical Management System API
  version: 1.0.0
  description: API documentation for School Medical Management System
servers:
  - url: http://localhost:3000/api
paths:
...
  /user/me:
    get:
      tags:
        - User Controller
      summary: Get user profile
      description: Retrieves a user's profile information
      operationId: getUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User retrieved successfully
                  data:
                    $ref: '#/components/schemas/User'
    put:
      tags:
        - User Controller
      summary: Update user profile
      description: Updates the profile information of a user
      operationId: editProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: 'johndoe'
                dob:
                  type: string
                  example: '31/12/1990'
                  description: Date of birth in DD/MM/YYYY format
                phone:
                  type: string
                  example: '+1234567890'
                gender:
                  type: string
                  enum: ['Male', 'Female']
              required:
                - username
                - dob
                - phone
                - gender
      responses:
        '200':
          description: User profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User profile updated successfully
                  data:
                    $ref: '#/components/schemas/User'
  /user/me/change-password:
    patch:
      tags:
        - User Controller
      summary: Change user password
      description: Updates the password of a user
      operationId: updatePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                  format: password
                  description: User's current password
                newPassword:
                  type: string
                  format: password
                  description: New password (minimum 6 characters)
                confirmNewPassword:
                  type: string
                  format: password
                  description: Confirmation of the new password
              required:
                - currentPassword
                - newPassword
                - confirmNewPassword
      responses:
        '200':
          description: Password updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Password updated successfully
...
components:
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          description: Unique identifier for the user
        username:
          type: string
          description: Username for authentication
        email:
          type: string
          description: User's email address
        role:
          type: string
          enum: [Parent, Admin, Teacher]
          description: User's role
        gender:
          type: string
          enum: ['Male', 'Female']
        dob:
          type: string
          format: date
          description: Date of birth for local authentication
        phone:
          type: string
          description: Phone number for local authentication
        isActive:
          type: boolean
          description: Indicates if the user account is active
        googleId:
          type: string
          nullable: true
          description: Google ID for users authenticated via Google
        authProvider:
          type: string
          enum: [local, google]
          description: Authentication provider used
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the user was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the user was last updated
      required:
        - _id
        - username
        - email
        - role
        - dob
        - phone
        - isActive
        - authProvider
        - googleId
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...